Parsing fixes, added several fixes to parsing
Fixed the parsing for each item, not a specific <td> item in the table is
accessed for each item and a regex is used to extract the data. If the item
is note found (ie. prof name in table is blank) then the entry is set to None,
the makes it easy to continue parsing without it crashing for keys that were
not created.

Fixed individual regexs to handle non-existant rooms/profs that were before being
recorded as existing, professors listed as TBA (to be annouced) are marked as None
as well as any online/virtual courses as there are no rooms for those.