MODIFIED: some instesting queries
Made some intresting queries with the data. Now we can find out which profs are teaching the most students and which are not teaching at all. Also, added a query that returns the course with the most students registered into it, calculus I. Finally, a few queries could not be completed with sql since capacity for web courses will be NULL, so programmatically it can be added to the number of students registered, or we could just not include courses that have null capacity in the count.