FIXED: a bug in get_rooms_open
Get rooms open was not checking if the time slot requested was overtop of a class time. Also, worked out the logic so now it will also catch the room as being taken if the class's start time == the requested start time or the class's end time == the requested end time.