FIXED: url error and parsing error
The url was not working because of a duplication of arguments. Also, the parsing was not working because there is a new class type that our regex is not capturing. A quick fix was done to it which is to set every unknown class type to class type 1. This is sufficent for now since the class type is not used extensively to find free rooms.