FIXED: radio button animation, book button
Radio button now properly displays its animation (its not cut off anymore). Also, if a table row is clicked the radio button is also clicked. Next, the book button has now been linked to whether a radio button is actually checked. If no radio button is checked then it will be disabled, and alternatively if a radio button is ckecked it will be enabled.