MODIFIED: login to actually check and add to db
Added login check the test username is specified in Loginactivity, and is username="database", password="test123". Use this to login and you wont have to do it again until the app's been destoryed. Also, added the current date as the default date for the search. Finally fixed a small db bug, with the getUser query retriving the username twice and not the password.