ADDED: preference dialog for deleting
Added a customized preference dialog which as the ability to register an on dialog closed listener. This helps handle the deletion of an account if the user confirms they wish to do so. Also, minor fixes were made to table layouts, 12 hour clock formating was taking up to much space. Seconds were dropped to support the date not wraping two lines.