FIXED: log out issues
Results and RoomsBooked did not explicitly check if the user was logged in. Now each tab activity checks and will launch a login activity if they see that the user isnt logged in. However this made duplicate logins (thanks android ViewPager). This was fixed by only launching checking if the user has logged in the onCreate.